Fixed TweakScaleCompanion_ReStockPlus def
Because things are never simple with AddOns, TweakScaleCompanion_ReStockPlus is actually a suite for 2 AddOns which are
- ReStockPlus
- RestockRigidLegs yes with no cap on the s of 'stock', double checked twice. 
Even RestockRigidLegs is an AddOn with only 3 parts defined, this causes this companion being a suite. I should wrote this in my resume lol.

Please check whether the & is accepted in your configuration file or not.